requirements.txt
# Requirements for running praelatus
amqp==2.1.4
billiard==3.5.0.2
celery==4.0.2
certifi==2017.4.17
chardet==3.0.4
Django==2.0.1
django-filter==1.0.4
django-guardian==1.4.8
django-redis==4.8.0
djangorestframework==3.7.7
gevent==1.2.1
greenlet==0.4.12
gunicorn==19.7.1
idna==2.5
jinja2==2.9.6
kombu==4.0.2
Markdown==2.6.8
MarkupSafe==1.0
olefile==0.44
Pillow==4.1.1
psycopg2==2.7.1
pytz==2017.2
PyYAML==3.12
redis==2.10.5
requests==2.17.3
six==1.10.0
urllib3==1.21.1
vine==1.1.3
ply==3.10
django-attachments==1.1.1
eventlet==0.21.0